chore: split resolvers
split resolvers in resolver and connectionResolver.
This makes it more explicity.
