+++ api/src/utils/index.js
@@ -1,0 +1,2 @@
+export { default as resolver } from './resolver';
+export { default as connectionResolver } from './connectionResolver';
